% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{add_description}
\usage{
add_description(adir = ".")
}
\arguments{
\item{adir}{a directory}
}
\value{
logical success
}
\description{
Adds a description file in CIP format
}
\author{
Reinhard Simon
}

