% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_tests}
\alias{add_tests}
\title{add_tests}
\usage{
add_tests(pkg = ".")
}
\arguments{
\item{pkg}{the path to the package}
}
\description{
Adds testthat files, examples and vignettes - all different means of
checking correctness.
}
\author{
Reinhard Simon
}

