% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ciptools-package.r, R/package.R
\docType{package}
\name{ciptools}
\alias{ciptools}
\alias{ciptools-package}
\title{ciptools.}
\description{
ciptools.

Tools to further standardize R development.
}
\author{
Reinhard Simon
}

