% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git_plus.R
\name{first_commit}
\alias{first_commit}
\title{first_commit}
\usage{
first_commit(pkg = ".", msg = "Initial commit.")
}
\arguments{
\item{pkg}{path to pkg}

\item{msg}{a message}
}
\description{
Does a first round of adding all files to git and pushing to github.
}
\author{
Reinhard Simon
}

