% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/format_r.R
\name{format_r}
\alias{format_r}
\title{format_r}
\usage{
format_r(adir = ".")
}
\arguments{
\item{adir}{a directory}
}
\value{
boolean true or false
}
\description{
formats R code
}
\author{
Reinhard Simon
}

