% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/has_description.R
\name{has_description}
\alias{has_description}
\title{has_description}
\usage{
has_description(dir = ".")
}
\arguments{
\item{dir}{a directory}
}
\value{
boolean true or false
}
\description{
Checks that the directory given has a DESCRIPTON file
}
\author{
Reinhard Simon
}

