% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/has_inst.R
\name{has_inst}
\alias{has_inst}
\title{has_inst}
\usage{
has_inst(adir = ".")
}
\arguments{
\item{adir}{a directory}
}
\value{
boolean true or false
}
\description{
Checks that the directory given has a subdirectory 'inst'
}
\author{
Reinhard Simon
}

