% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/has_r.R
\name{has_r}
\alias{has_r}
\title{has_r}
\usage{
has_r(dir = ".")
}
\arguments{
\item{dir}{a directory}
}
\value{
boolean true or false
}
\description{
Checks that the directory given has an R subdirectory
}
\author{
Reinhard Simon
}

