% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{new_description}
\alias{new_description}
\title{new_description}
\usage{
new_description(adir = ".", title = "My Productivity Tool (Use Title Case)",
  description = "Be a bit more specific", name_first = "First",
  name_last = "Last", email = "first.last@email.com", role = c("aut",
  "cre"), copyright = "International Potato Center",
  license = "MIT + file LICENSE")
}
\arguments{
\item{adir}{a directory}

\item{title}{for DESCRIPTION and README files}

\item{description}{for DESCRIPTION and README files}

\item{name_first}{of author}

\item{name_last}{of author}

\item{email}{reference email. Must match.}

\item{copyright}{Copyright holder.}

\item{license}{Standard license.}
}
\value{
logical success
}
\description{
Adds a description file in CIP format
}
\author{
Reinhard Simon
}

