% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{replace_description}
\alias{replace_description}
\title{Replace a description}
\usage{
replace_description(field, name, pkg = ".")
}
\arguments{
\item{field}{the field name}

\item{name}{the content of the field}

\item{pkg}{the package directory}
}
\description{
Replace a field content in he description file.
}

