% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{use_cip}
\alias{use_cip}
\title{use_cip}
\usage{
use_cip(pkg = ".", persons = c(person("Reinard", "Simon", ,
  "r.simon@cgiar.org", role = c("auth", "cre"))), title = "A Good Title",
  description = "A thorough description")
}
\arguments{
\item{pkg}{path to package}

\item{persons}{a vector of persons}

\item{title}{a title}

\item{description}{a description}
}
\description{
Add CIP standards
}
\author{
Reinhard Simon
}

