% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{use_citation}
\alias{use_citation}
\title{use_citation}
\usage{
use_citation(pkg = ".")
}
\arguments{
\item{pkg}{the package directory}
}
\value{
boolean
}
\description{
Adds a CITATION file to the inst sub-directory. The template
creates a citation similary to the auto-generated one. However,
this file may be customized.
}
\author{
Reinhard Simon
}

