% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_coveralls}
\alias{use_coveralls}
\title{Use coveralls}
\usage{
use_coveralls(pkg = ".")
}
\arguments{
\item{pkg}{a directory}
}
\description{
Taken from Jim Heesters modified devtools package
}

