% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{Add to description}
\usage{
add_description(field = "Package", name = basename(getwd()), pkg = ".")
}
\arguments{
\item{field}{the field name}

\item{name}{the content of the field}

\item{pkg}{the package directory}
}
\description{
Add a field to the description file. This function is a wrappe around a function
from the devtools package.
}

