% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_function.R
\name{add_function}
\alias{add_function}
\title{add_function}
\usage{
add_function(title)
}
\arguments{
\item{title}{a name for the function}
}
\description{
Adds a new user visible function
}
\author{
Reinhard Simon
}
\seealso{
Other convenience: \code{\link{build}}
}

