% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_description.R
\name{new_description}
\alias{new_description}
\title{new_description}
\usage{
new_description(adir = ".", title = "My Productivity Tool (Use Title Case)",
  description = "Be a bit more specific!", persons = c(person("First",
  "Last", "M", "first.last@email.com", role = c("aut", "cre"))),
  copyright = "International Potato Center", license = "MIT + file LICENSE")
}
\arguments{
\item{adir}{a directory}

\item{title}{for DESCRIPTION and README files}

\item{description}{for DESCRIPTION and README files}

\item{persons}{a vector of persons}

\item{copyright}{Copyright holder.}

\item{license}{Standard license.}
}
\value{
logical success
}
\description{
Adds a description file in CIP format
}
\author{
Reinhard Simon
}

